+++ python/sglang/srt/passes/__init__.py
@@ -1,0 +1,7 @@
+from .base_pass import BaseFXPass
+from .silu_mul_fusion_pass import SiLUMulFusionPass
+
+__all__ = [
+    "BaseFXPass",
+    "SiLUMulFusionPass",
+]
